requirements.txt
google-adk
litellm
fastapi
uvicorn
httpx
pydantic
openai
streamlit
pyyaml
elasticsearch==8.14
mcp-server